pyproject.toml
[tool.poetry]
name = "Difference calculator"
version = "0.1.1"
description = ""
authors = ["Evgeny <litun-07@mail.ru>"]
packages = [
	{ include = "difference_calculator" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "difference_calculator.scripts.gendiff:main"
